Restructuring envelope and parts with c code
