+++ gmime/gmime.go
@@ -2,69 +2,9 @@ package gmime
 
 /*
 #cgo pkg-config: gmime-3.0
-#include <stdlib.h>
-#include <strings.h>
-#include <gmime/gmime.h>
-
-static GMimeMessage *gmime_parse (const char *buffer, size_t len) {
-	GMimeStream *stream = g_mime_stream_mem_new_with_buffer (buffer, len);
-	GMimeParser *parser = g_mime_parser_new_with_stream (stream);
-	g_object_unref (stream);
-	GMimeMessage *message = g_mime_parser_construct_message (parser, NULL);
-	g_object_unref (parser);
-
-	InternetAddressList *list = g_mime_message_get_addresses (message, GMIME_ADDRESS_TYPE_TO);
-	int listLen = internet_address_list_length (list);
-	for(int i = 0; i < listLen; i++) {
-		InternetAddress *addr = internet_address_list_get_address (list, i);
-		printf("Name: %s\n", internet_address_get_name (addr));
-		printf("Address: %s\n", internet_address_mailbox_get_addr ((InternetAddressMailbox *)addr));
-	}
-
-	return message;
-}
-
-static char* gmime_get_content_type_string (GMimeObject *object) {
-	GMimeContentType *ctype = g_mime_object_get_content_type (object);
-	return g_mime_content_type_get_mime_type (ctype);
-}
-
-static char* gmime_get_content_string (GMimeObject *object) {
-	if (!GMIME_IS_TEXT_PART (object)) {
-		return NULL;
-	}
-	return g_mime_text_part_get_text ((GMimeTextPart *) object);
-}
-
-static gboolean gmime_is_text_part (GMimeObject *object) {
-	return GMIME_IS_TEXT_PART (object);
-}
-
-static GByteArray *gmime_get_bytes (GMimeObject *object) {
-	GMimeStream *stream;
-	GMimeDataWrapper *content;
-	GByteArray *buf;
-
-	if (!(content = g_mime_part_get_content ((GMimePart *) object)))
-		return NULL;
-	stream = g_mime_stream_mem_new ();
-	ssize_t size = g_mime_data_wrapper_write_to_stream (content, stream);
-	g_mime_stream_flush (stream);
-
-	buf = g_mime_stream_mem_get_byte_array ((GMimeStreamMem *) stream);
-	g_mime_stream_mem_set_owner ((GMimeStreamMem *) stream, FALSE);
-
-	g_object_unref (stream);
-	return buf;
-}
-
+#include "gmime.h"
 */
 import "C"
-import (
-	"errors"
-	"net/textproto"
-	"unsafe"
-)
 
 // This function call automatically by runtime
 func init() {
@@ -92,128 +32,4 @@ func gobool(b C.gboolean) bool {
 // free up memory
 func unref(referee C.gpointer) {
 	C.g_object_unref(referee)
-}
-
-// Envelope wraps gmime message object and has methods to access it
-type Envelope struct {
-	gmimeMessage *C.GMimeMessage
-}
-
-// Part is a wrapper for message parts
-type Part struct {
-	gmimePart *C.GMimeObject
-}
-
-// ContentType returns part's content type
-func (p *Part) ContentType() string {
-	ctype := C.gmime_get_content_type_string(p.gmimePart)
-	defer C.g_free(C.gpointer(unsafe.Pointer(ctype)))
-	return C.GoString(ctype)
-}
-
-// IsText returns true if part's mime is text/*
-func (p *Part) IsText() bool {
-	return gobool(C.gmime_is_text_part(p.gmimePart))
-}
-
-// Text returns text portion of the part if it's mime is text/*
-func (p *Part) Text() string {
-	content := C.gmime_get_content_string(p.gmimePart)
-	defer C.g_free(C.gpointer(unsafe.Pointer(content)))
-	return C.GoString(content)
-}
-
-// Bytes returns decoded raw bytes of the part, most useful to access attachment data
-func (p *Part) Bytes() []byte {
-	b := C.gmime_get_bytes(p.gmimePart)
-	defer C.g_byte_array_free((*C.GByteArray)(unsafe.Pointer(b)), C.TRUE)
-	return C.GoBytes(unsafe.Pointer(b.data), C.int(b.len))
-}
-
-// SetText replaces text content if part is text/*
-func (p *Part) SetText(text string) error {
-	// TODO: Optimize this
-	cstr := C.CString(text)
-	defer C.free(unsafe.Pointer(cstr))
-	C.g_mime_text_part_set_text((*C.GMimeTextPart)(unsafe.Pointer(p.gmimePart)), cstr)
-	return nil
-}
-
-func (p *Part) Headers() textproto.MIMEHeader {
-	return nil
-}
-
-// Parse parses message and returns Message
-func Parse(data string) *Envelope {
-	// very inefficient
-	cBuf := C.CString(data)
-	defer C.free(unsafe.Pointer(cBuf))
-	return &Envelope{
-		gmimeMessage: C.gmime_parse(cBuf, C.size_t(len(data))),
-	}
-}
-
-// Subject returns envelope's Subject
-func (m *Envelope) Subject() string {
-	subject := C.g_mime_message_get_subject(m.gmimeMessage)
-	return C.GoString(subject)
-}
-
-// SetSubject returns envelope's Subject
-func (m *Envelope) SetSubject(subject string) {
-}
-
-func (m *Envelope) Headers() textproto.MIMEHeader {
-	return nil
-}
-
-func (m *Envelope) SetHeader() textproto.MIMEHeader {
-	return nil
-}
-
-func (m *Envelope) Header(header string) []string {
-	return nil
-}
-
-// ContentType returns envelope's content-type
-func (m *Envelope) ContentType() string {
-	mimePart := C.g_mime_message_get_mime_part(m.gmimeMessage)
-	ctype := C.gmime_get_content_type_string(mimePart)
-	defer C.g_free(C.gpointer(unsafe.Pointer(ctype)))
-	return C.GoString(ctype)
-}
-
-// Walk iterates all message parts and executes callback on each part
-func (m *Envelope) Walk(cb func(p *Part)) {
-	partIter := C.g_mime_part_iter_new((*C.GMimeObject)(unsafe.Pointer(m.gmimeMessage)))
-	for {
-		next := C.g_mime_part_iter_next(partIter)
-		if !gobool(next) {
-			break
-		}
-		currentPart := C.g_mime_part_iter_get_current(partIter)
-		part := &Part{
-			gmimePart: currentPart,
-		}
-		cb(part)
-	}
-}
-
-// Export composes mime from envelope
-func (m *Envelope) Export() ([]byte, error) {
-	// TODO: optimize this, bundle cgo calls
-	stream := C.g_mime_stream_mem_new()                        // need unref
-	defer C.g_object_unref(C.gpointer(unsafe.Pointer(stream))) // unref
-	nWritten := C.g_mime_object_write_to_stream((*C.GMimeObject)(unsafe.Pointer(m.gmimeMessage)), nil, stream)
-	if nWritten <= 0 {
-		return nil, errors.New("can't write to stream")
-	}
-	// byteArray is owned by stream and will be freed with it
-	byteArray := C.g_mime_stream_mem_get_byte_array((*C.GMimeStreamMem)(unsafe.Pointer(stream)))
-	return C.GoBytes(unsafe.Pointer(byteArray.data), (C.int)(nWritten)), nil
-}
-
-// Close frees up message resources
-func (m *Envelope) Close() {
-	C.g_object_unref(C.gpointer(m.gmimeMessage))
 }
